Add files via upload
Star Fitness Club wants to calculate the Body Mass Index (BMI) for all its customers. Customers may enter their height and weight. Each customer should know their body mass index and weight loss or gain goals to be fit and healthy. Help them to achieve this with the help of a Java program. 

The formula for calculating BMI is weight/((height/100) *(height/100)).

Display the bmi with respect to 2 decimal points.

If bmi is greater than or equal to 25, print "You are overweight" and then print the number of kilograms to be reduced to become fit as "Reduce <<kgs>>kg to be fit".

If bmi is less than 25 and greater than or equal to 18.5, print "You are fit and healthy".

If bmi is less than 18.5, print "You are underweight" and then print the number of kilograms to be gained to become fit as "Gain <<kgs>>kg to be fit".


Note:

-In the Sample Input / Output provided, the highlighted text in bold corresponds to the input given by the user, and the rest of the text represents the output.
-Ensure to follow the object-oriented specifications provided in the question description.
-Ensure to provide the names for classes, attributes, and methods as specified in the question description.
-Adhere to the code template, if provided.
-Please do not use System.exit(0) to terminate the program