Description
Johanna, a 5-year-old girl, is learning the number names. She is having difficulty remembering all of the words. To give her extra practice, her mother assigns a number and instructs her to write the number names separated by white space. Help her mum by writing a Java program that prints the number names of the numbers so Johanna may check the answers herself.

Assume that she is learning from zero to nine.


Note:

Number should not start with zero
In the Sample Input / Output provided, the highlighted text in bold corresponds to the input given by the user, and the rest of the text represents the output.
Ensure to follow the object-oriented specifications provided in the question description.
Ensure to provide the names for classes, attributes, and methods as specified in the question description.
Adhere to the code template, if provided.