Description
Henry has a magic goose. The goose will lay eggs only for 30 consecutive days in a year. The number of eggs laid by the goose will be in the following pattern. On day 1, it will lay 1 egg, on day 2, it will lay 2 eggs, on day 3 it will lay 5 eggs, on day 4 12 eggs and so on. Write a java program to calculate the number of eggs the goose will lay on a particular day. 

Note: The egg laying pattern follows the pell series. i.e.,[1,2,5,12,29,70...]

If the day is more than 30 or negative or zero, then print "<<day>> is an invalid day". 


Note:

In the Sample Input / Output provided, the highlighted text in bold corresponds to the input given by the user, and the rest of the text represents the output.
Ensure to follow the object-oriented specifications provided in the question description.
Ensure to provide the names for classes, attributes, and methods as specified in the question description.
Adhere to the code template, if provided.