Description
awan owns a lottery ticket. He can win the lottery only if the unlucky code is not present in the ticket id.  

Requirements:

Number of times of unlucky code in the ticket Id decide the winner.
If there is no unlucky code in the ticket Id, then print as "<<Ticket Id>> is lucky ticket".
If the count of the unlucky code is less than 3, then print as "<<Ticket Id>> is partially lucky".
If the count of the unlucky code is greater than or equal to 3, then print as "<<Ticket Id>> is unlucky ticket".
Note:

In the Sample Input / Output provided, the highlighted text in bold corresponds to the input given by the user, and the rest of the text represents the output.
Ensure to follow the object-oriented specifications provided in the question description.
Ensure to provide the names for classes, attributes, and methods as specified in the question description.
Adhere to the code template, if provided.